Changes to be committed: 	new file:   src/diagnostics.ts 	modified:   src/extension.ts
